requirements.txt
torch==1.8.1
torchvision==0.9.1
numpy==1.20
pillow==8.2.0
tensorboard
sklearn
tqdm
pyyaml
easydict
termcolor